services/ai/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
prometheus-fastapi-instrumentator>=7.0.0